+++ lib/widgets/journal/entry_details/entry_detail_header.dart
@@ -34,7 +34,6 @@ class EntryDetailHeader extends StatelessWidget {
           children: [
             Row(
               children: [
-                const SizedBox(width: 5),
                 SwitchIconWidget(
                   tooltip: localizations.journalFavoriteTooltip,
                   onPressed: cubit.toggleStarred,