+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Upgrade dependencies
+- Whitespace on entry detail page
+
+### Fixed:
+- Styling of DateTime modal bottom sheet
+
+## [0.8.296] - 2023-03-30
 ### Fixed:
 - Habit completion chart display range issue